# api/openapi.yaml
openapi: 3.1.0
info:
  title: model-uptime HTTP API
  version: 0.1.0
  description: |
    Contract for the public status endpoints and the authenticated administration API.
    The web UI and API are served from the same origin by the model-uptime binary.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/heatmap:
    get:
      operationId: getHeatmap
      summary: Read aggregated service health history
      security: []
      parameters:
        - name: range
          in: query
          required: false
          schema:
            type: string
            enum: [1d, 7d, 30d, day, week, month]
            default: 7d
      responses:
        '200':
          description: Health heatmap for the requested range.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HeatmapResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/admin/setup:
    post:
      operationId: setupAdminPassword
      summary: Set the initial administrator password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupTokenRequest'
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/admin/login:
    post:
      operationId: loginAdmin
      summary: Validate an administrator password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginTokenRequest'
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/services:
    get:
      operationId: listServices
      summary: List configured monitoring services
      responses:
        '200':
          description: Configured services with API keys masked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createService
      summary: Create a monitoring service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceInput'
      responses:
        '201':
          description: Created service with its API key masked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
    patch:
      operationId: updateServices
      summary: Apply supported fields to multiple monitoring services
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkServicePatchRequest'
      responses:
        '200':
          description: Updated services with API keys masked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/services/{id}:
    parameters:
      - $ref: '#/components/parameters/ServiceID'
    put:
      operationId: updateService
      summary: Replace a monitoring service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceInput'
      responses:
        '200':
          description: Updated service with its API key masked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteService
      summary: Delete a monitoring service
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/services/{id}/test:
    parameters:
      - $ref: '#/components/parameters/ServiceID'
    post:
      operationId: testService
      summary: Probe a monitoring service immediately
      responses:
        '200':
          description: Probe result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProbeResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/admin/services/{id}/duplicate:
    parameters:
      - $ref: '#/components/parameters/ServiceID'
    post:
      operationId: duplicateService
      summary: Duplicate a monitoring service
      responses:
        '201':
          description: Duplicated service with its API key masked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/admin/page:
    get:
      operationId: getPageConfig
      summary: Read public page configuration
      responses:
        '200':
          description: Current public page configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      operationId: updatePageConfig
      summary: Replace public page configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PageConfigInput'
      responses:
        '200':
          description: Normalized public page configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/telegram:
    get:
      operationId: getTelegramConfig
      summary: Read Telegram notification configuration
      responses:
        '200':
          description: Telegram configuration with the bot token masked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramView'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      operationId: updateTelegramConfig
      summary: Replace Telegram notification configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramConfig'
      responses:
        '200':
          description: Normalized Telegram configuration with the bot token masked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/admin/telegram/test:
    post:
      operationId: testTelegramNotification
      summary: Send an event or daily test notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramTestRequest'
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
  /api/admin/update:
    get:
      operationId: getUpdateStatus
      summary: Read release and updater status
      responses:
        '200':
          description: Current release and updater status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    post:
      operationId: startUpdate
      summary: Trigger an update to the latest stable release
      responses:
        '202':
          description: Update accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/admin/update/check:
    post:
      operationId: checkForUpdate
      summary: Force a stable release and image availability check
      responses:
        '200':
          description: Current release and updater status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Administrator password returned to the UI session after successful validation.
  parameters:
    ServiceID:
      name: id
      in: path
      required: true
      description: Stable monitoring service identifier.
      schema:
        type: string
        minLength: 1
  responses:
    OK:
      description: Operation completed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OKResponse'
    BadRequest:
      description: Invalid request body, parameter, or configuration.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid administrator bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with the current application state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadGateway:
      description: An upstream service failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: The requested application service is not configured.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal application failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
    OKResponse:
      type: object
      additionalProperties: false
      required: [ok]
      properties:
        ok:
          type: boolean
          const: true
    LoginTokenRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
    SetupTokenRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 8
...
    ServiceInput:
      type: object
      additionalProperties: false
      required: [name, protocol, base_url]
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          type: string
        protocol:
          type: string
          enum: [chat, response, message, http]
        model:
          type: string
        base_url:
          type: string
          format: uri
        api_key:
          type: string
          description: Masked in responses; use the unchanged sentinel or omit it to retain the stored value.
        path:
          type: string
        sort_order:
          type: integer
          minimum: 0
        interval_sec:
          type: integer
          minimum: 5
        timeout_sec:
          type: integer
          minimum: 1
          maximum: 300
        warning_sec:
          type: integer
          minimum: 1
          maximum: 300
        enabled:
          type: boolean
        stream:
          type: boolean
        method:
          type: string
        headers:
          type: object
          additionalProperties:
            type: string
        body:
          type: string
        expect_status:
          type: integer
          minimum: 100
          maximum: 599
    Service:
      allOf:
        - $ref: '#/components/schemas/ServiceInput'
        - type: object
          required: [id]
    ServiceEnvelope:
      type: object
      additionalProperties: false
      required: [service]
      properties:
        service:
          $ref: '#/components/schemas/Service'
    ServiceList:
      type: object
      additionalProperties: false
      required: [services]
      properties:
        services:
          type: array
          items:
            $ref: '#/components/schemas/Service'
    ServicePatch:
      type: object
      additionalProperties: false
      properties:
        enabled:
          type: boolean
        interval_sec:
          type: integer
          minimum: 5
        timeout_sec:
          type: integer
          minimum: 1
          maximum: 300
        warning_sec:
          type: integer
          minimum: 1
          maximum: 300
        stream:
          type: boolean
    BulkServicePatchRequest:
      type: object
      additionalProperties: false
      required: [ids, patch]
      properties:
        ids:
          type: array
          minItems: 1
          items:
            type: string
        patch:
          $ref: '#/components/schemas/ServicePatch'
    PageConfig:
      type: object
      additionalProperties: false
      required:
        - title
        - subtitle
        - probe_comment
        - public_url
        - history_len
        - refresh_sec
        - enable_command_animation
        - show_uptime
        - show_samples
        - show_latency
        - show_avg_load
      properties:
        title:
          type: string
        subtitle:
          type: string
        probe_comment:
          type: string
        public_url:
          type: string
        history_len:
          type: integer
          minimum: 1
          maximum: 200
        refresh_sec:
          type: integer
          minimum: 1
          maximum: 60
        enable_command_animation:
          type: boolean
        show_uptime:
          type: boolean
        show_samples:
          type: boolean
        show_latency:
          type: boolean
        show_avg_load:
          type: boolean
    PageConfigInput:
      type: object
      additionalProperties: false
      description: Partial input is accepted; omitted fields are filled with server defaults.
      properties:
        title:
          type: string
        subtitle:
          type: string
        probe_comment:
          type: string
        public_url:
          type: string
        history_len:
          type: integer
          minimum: 1
          maximum: 200
        refresh_sec:
          type: integer
          minimum: 1
          maximum: 60
        enable_command_animation:
          type: boolean
        show_uptime:
          type: boolean
        show_samples:
          type: boolean
        show_latency:
          type: boolean
        show_avg_load:
          type: boolean
    TelegramSubscription:
      type: object
      additionalProperties: false
      required: [id, name, enabled, chat_id, language, service_ids, template]
      properties:
        id:
          type: string
        name:
          type: string
        enabled:
          type: boolean
        chat_id:
          type: string
        language:
          type: string
          enum: [zh-CN, en-US]
        service_ids:
          type: array
          items:
            type: string
        template:
          type: string
    TelegramConfig:
      type: object
      additionalProperties: false
      properties:
        bot_token:
          type: string
          description: Use the masked sentinel to retain the stored token.
        subscriptions:
          type: array
          items:
            $ref: '#/components/schemas/TelegramSubscription'
    TelegramView:
      type: object
      additionalProperties: false
      required: [bot_token, token_configured, subscriptions, templates]
      properties:
        bot_token:
          type: string
        token_configured:
          type: boolean
        subscriptions:
          type: array
          items:
            $ref: '#/components/schemas/TelegramSubscription'
        templates:
          type: object
          additionalProperties: false
          required: [zh, en]
          properties:
            zh:
              type: string
            en:
              type: string
    TelegramTestRequest:
      type: object
      additionalProperties: false
      required: [subscription_id]
      properties:
        subscription_id:
          type: string
        kind:
          type: string
          enum: [event, daily]
          default: event
    ProbeResult:
      type: object
      additionalProperties: false
      required: [ok, ts, latency_ms]
      properties:
        ok:
          type: boolean
        ts:
          type: integer
          format: int64
        started_at:
          type: integer
          format: int64
        latency_ms:
          type: integer
          format: int64
        error:
          type: string
...
    HeatmapCell:
      type: object
      additionalProperties: false
      required:
        - start_ts
        - end_ts
        - status
        - intensity
        - coverage_pct
        - actual_samples
        - expected_samples
        - healthy_samples
        - warning_samples
        - failed_samples
        - uptime_pct
        - avg_latency_ms
        - p95_latency_ms
      properties:
        start_ts:
          type: integer
          format: int64
        end_ts:
          type: integer
          format: int64
        status:
          type: string
          enum: [healthy, warning, failing, unobserved, insufficient]
        intensity:
          type: integer
        coverage_pct:
          type: number
        actual_samples:
          type: integer
        expected_samples:
          type: integer
        healthy_samples:
          type: integer
        warning_samples:
          type: integer
        failed_samples:
          type: integer
        uptime_pct:
          type: number
        avg_latency_ms:
          type: integer
          format: int64
        p95_latency_ms:
          type: integer
          format: int64
    HeatmapService:
      type: object
      additionalProperties: false
      required:
        - id
        - model
        - status
        - samples
        - latency_samples
        - uptime_pct
        - p95_latency_ms
        - cells
      properties:
        id:
          type: string
        model:
          type: string
        provider:
          type: string
        status:
          type: string
          enum: [healthy, warning, failing, unobserved, insufficient]
        samples:
          type: integer
        latency_samples:
          type: integer
        uptime_pct:
          type: number
        p95_latency_ms:
          type: integer
          format: int64
        cells:
          type: array
          items:
            $ref: '#/components/schemas/HeatmapCell'
    HeatmapResponse:
      type: object
      additionalProperties: false
      required: [generated_at, range, timezone, bucket_sec, rows, columns, services]
      properties:
        generated_at:
          type: integer
          format: int64
        range:
          type: string
          enum: [1d, 7d, 30d]
        timezone:
          type: string
          const: Asia/Shanghai
        bucket_sec:
          type: integer
          format: int64
        rows:
          type: array
          items:
            type: string
        columns:
          type: array
          items:
            type: string
        services:
          type: array
          items:
            $ref: '#/components/schemas/HeatmapService'
    UpdateStatus:
      type: object
      additionalProperties: false
      required:
        - enabled
        - current_version
        - update_available
        - deployment_tag
        - updating
      properties:
        enabled:
          type: boolean
        disabled_reason:
          type: string
        current_version:
          type: string
        latest_version:
          type: string
        update_available:
          type: boolean
        deployment_tag:
          type: string
        commit:
          type: string
        built_at:
          type: string
        checked_at:
          type: string
          format: date-time
        updating:
          type: boolean
        last_update_error:
          type: string
    UpdateAccepted:
      type: object
      additionalProperties: false
      required: [ok, target_version]
      properties:
        ok:
          type: boolean
          const: true
        target_version:
          type: string